Move creation of cloud context into its own flight and wait for it to complete (WOR-1546).
