+++ service/src/main/java/bio/terra/workspace/service/resource/controlled/flight/clone/workspace/CreateCloudContextIdsForFutureStepsStep.java
@@ -13,20 +13,21 @@ import bio.terra.workspace.service.workspace.flight.WorkspaceFlightMapKeys.Contr
  * Generate flight IDs ahead of the steps that need them. The steps can then check the existence of
  * the flights or workspaces before creation to ensure idempotency.
  */
-public class CreateIdsForFutureStepsStep implements Step {
+public class CreateCloudContextIdsForFutureStepsStep implements Step {
 
   @Override
   public StepResult doStep(FlightContext context) throws InterruptedException, RetryException {
     FlightMap workingMap = context.getWorkingMap();
     workingMap.put(
-        ControlledResourceKeys.WORKSPACE_CREATE_FLIGHT_ID, context.getStairway().createFlightId());
+        ControlledResourceKeys.CREATE_GCP_CLOUD_CONTEXT_FLIGHT_ID,
+        context.getStairway().createFlightId());
     workingMap.put(
-        ControlledResourceKeys.CLONE_ALL_RESOURCES_FLIGHT_ID,
+        ControlledResourceKeys.CREATE_AZURE_CLOUD_CONTEXT_FLIGHT_ID,
         context.getStairway().createFlightId());
     validateRequiredEntries(
         workingMap,
-        ControlledResourceKeys.WORKSPACE_CREATE_FLIGHT_ID,
-        ControlledResourceKeys.CLONE_ALL_RESOURCES_FLIGHT_ID);
+        ControlledResourceKeys.CREATE_GCP_CLOUD_CONTEXT_FLIGHT_ID,
+        ControlledResourceKeys.CREATE_AZURE_CLOUD_CONTEXT_FLIGHT_ID);
     return StepResult.getStepResultSuccess();
   }
